Fixed printf format string and argument types for 64 bit builds.
darcs-hash:20060510120714-ced27-71a8b92d9b7f9b09cb1957555a2ba79329e01784.gz

+++ lib/tre-compile.c
@@ -1932,7 +1932,7 @@ tre_compile(regex_t *preg, const tre_char_t *regex, size_t n, int cflags)
   parse_ctx.len = n;
   parse_ctx.cflags = cflags;
   parse_ctx.max_backref = -1;
-  DPRINT(("tre_compile: parsing '%.*" STRF "'\n", n, regex));
+  DPRINT(("tre_compile: parsing '%.*" STRF "'\n", (int)n, regex));
   errcode = tre_parse(&parse_ctx);
   if (errcode != REG_OK)
     ERROR_EXIT(errcode);
